# Conflicts:
#	src/Main.java

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {100,200,300,2000,400,200};
+        long[] sales = {100,200,300,2000,400,200};
         SalesManager manager = new SalesManager(sales);
 
         System.out.println(manager.max());
